+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add option for nearest when scaling image.
 - Add codes related to mouse events
 - Parse mouse events (classic, SGR, UTF-8, URXVT)
+- Parse focus events
 
 ## v1.2.2
 - Fix `writemcln` macro.